Node: Rename Arbitrum connector (#2432)
* Node: Rename Arbitrum connector


* Rename the class
