+++ node/pkg/watchers/evm/watcher.go
@@ -272,7 +272,8 @@ func (w *Watcher) Run(ctx context.Context) error {
 			p2p.DefaultRegistry.AddErrorCount(w.chainID, 1)
 			return fmt.Errorf("creating block poll connector failed: %w", err)
 		}
-		w.ethConn, err = connectors.NewArbitrumConnector(ctx, pollConnector)
+		// The standard geth BlockByHash() does not work on Arbitrum.
+		w.ethConn, err = connectors.NewConnectorWithGetTimeOfBlockOverride(ctx, pollConnector)
 		if err != nil {
 			ethConnectionErrors.WithLabelValues(w.networkName, "dial_error").Inc()
 			p2p.DefaultRegistry.AddErrorCount(w.chainID, 1)
@@ -318,9 +319,8 @@ func (w *Watcher) Run(ctx context.Context) error {
 				p2p.DefaultRegistry.AddErrorCount(w.chainID, 1)
 				return fmt.Errorf("creating block poll connector failed: %w", err)
 			}
-			// I know this says Arbitrum.  That's just what the type is called.
-			// But we need it the TimeOfBlockByHash() implementation.
-			w.ethConn, err = connectors.NewArbitrumConnector(ctx, pollConnector)
+			// The standard geth BlockByHash() does not work on Optimism Bedrock.
+			w.ethConn, err = connectors.NewConnectorWithGetTimeOfBlockOverride(ctx, pollConnector)
 			if err != nil {
 				ethConnectionErrors.WithLabelValues(w.networkName, "dial_error").Inc()
 				p2p.DefaultRegistry.AddErrorCount(w.chainID, 1)
@@ -357,8 +357,8 @@ func (w *Watcher) Run(ctx context.Context) error {
 			p2p.DefaultRegistry.AddErrorCount(w.chainID, 1)
 			return fmt.Errorf("creating block poll connector failed: %w", err)
 		}
-		// Use the Arbitrum connector to get the TimeOfBlockByHash() implementation.
-		w.ethConn, err = connectors.NewArbitrumConnector(ctx, pollConnector)
+		// The standard geth BlockByHash() does not work on Base.
+		w.ethConn, err = connectors.NewConnectorWithGetTimeOfBlockOverride(ctx, pollConnector)
 		if err != nil {
 			ethConnectionErrors.WithLabelValues(w.networkName, "dial_error").Inc()
 			p2p.DefaultRegistry.AddErrorCount(w.chainID, 1)
